✨ feat(api): delete user can only current user or admin

+++ apps/api/src/modules/user/user.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common"
+import { ForbiddenException, Injectable } from "@nestjs/common"
 
 import {
   DeleteUserDto,
@@ -6,6 +6,7 @@ import {
   FindOneUserByIdDto,
 } from "./dto/user.dto"
 import { UserRepository } from "./user.repository"
+import { JwtPayload } from "../../core/interfaces"
 
 @Injectable()
 export class UserService {
@@ -23,7 +24,11 @@ export class UserService {
     return await this.userRepository.findOneByEmailWithPassword(email)
   }
 
-  async delete({ id }: DeleteUserDto) {
+  async delete({ id }: DeleteUserDto, user: JwtPayload) {
+    if (user.sub !== id && user.role !== "admin") {
+      throw new ForbiddenException()
+    }
+
     const [deletedUser] = await this.userRepository.deleteOne(id)
 
     return deletedUser
